refactor(pkg): move actions within web-pkg
