+++ packages/web-app-files/src/composables/actions/files/useFileActionsOpenShortcut.ts
@@ -48,6 +48,7 @@ export const useFileActionsOpenShortcut = () => {
       name: 'open-shortcut',
       icon: 'external-link',
       category: 'primary',
+      hasPriority: true,
       handler,
       label: () => {
         return $gettext('Open shortcut')
