+++ packages/web-pkg/src/composables/actions/files/index.ts
@@ -1,8 +1,5 @@
-export * from './useFileActions'
 export * from './useFileActionsDelete'
 export * from './useFileActionsDownloadFile'
-export * from './useFileActionsOpenWithDefault'
-export * from './useFileActionsOpenWithApp'
 export * from './useFileActionsRestore'
 export * from './useFileActionsSaveAs'
 export * from './useFileActionsUndoDelete'
