+++ packages/web-pkg/src/composables/actions/helpers/index.ts
@@ -1,5 +1,1 @@
 export * from './useFileActionsDeleteResources'
-export * from './useIsAppActive'
-export * from './useIsFilesAppActive'
-export * from './useIsResourceNameValid'
-export * from './useIsSearchActive'
