--- packages/web-pkg/src/composables/actions/helpers/useIsAppActive.ts
@@ -1,7 +1,0 @@
-import { computed, unref } from 'vue'
-import { useRoute } from '../../router'
-
-export const useIsAppActive = () => {
-  const currentRoute = useRoute()
-  return computed(() => unref(currentRoute)?.query?.contextRouteName)
-}
