+++ packages/web-pkg/src/composables/actions/index.ts
@@ -3,4 +3,7 @@ export * from './files'
 export * from './spaces'
 export * from './types'
 
+export * from './useFileActions'
+export * from './useFileActionsOpenWithDefault'
+export * from './useFileActionsOpenWithApp'
 export * from './useWindowOpen'
