+++ packages/web-pkg/src/composables/actions/useFileActionsOpenWithApp.ts
@@ -1,15 +1,15 @@
-import { FileAction, FileActionOptions } from '../types'
+import { FileAction, FileActionOptions } from './types'
 import { computed, markRaw, unref } from 'vue'
 import { useGettext } from 'vue3-gettext'
-import { useAppsStore, useModals } from '../../piniaStores'
+import { useAppsStore, useModals } from '../piniaStores'
 import { storeToRefs } from 'pinia'
-import FilePickerModal from '../../../components/Modals/FilePickerModal.vue'
-import { useFolderLink } from '../../folderLink'
-import { useIsFilesAppActive } from '../helpers'
+import FilePickerModal from '../../components/Modals/FilePickerModal.vue'
+import { useFolderLink } from '../folderLink'
+import { useIsFilesAppActive } from '../appDefaults'
 import { isShareSpaceResource, Resource } from '@opencloud-eu/web-client'
-import { useGetMatchingSpace } from '../../spaces'
+import { useGetMatchingSpace } from '../spaces'
 import { useFileActions } from './useFileActions'
-import { useRouter } from '../../router'
+import { useRouter } from '../router'
 import { LocationQuery } from 'vue-router'
 
 export const useFileActionsOpenWithApp = ({ appId }: { appId: string }) => {
