+++ packages/web-pkg/src/composables/appDefaults/index.ts
@@ -6,3 +6,4 @@ export * from './useAppFolderHandling'
 export * from './useAppMeta'
 export * from './useAppNavigation'
 export * from './useDocumentTitle'
+export * from './useIsAppActive'
