+++ packages/web-pkg/src/composables/resources/index.ts
@@ -3,6 +3,7 @@ export * from './useCanListVersions'
 export * from './useGetResourceContext'
 export * from './useLinkTargetRoute'
 export * from './useLoadPreview'
+export * from './useIsResourceNameValid'
 export * from './useResourceContents'
 export * from './useResourceIndicators'
 export * from './useResourceLink'
