+++ packages/web-pkg/src/composables/search/index.ts
@@ -1,2 +1,3 @@
 export * from './constants'
+export * from './useIsSearchActive'
 export * from './useSearch'
